Move serialization logic from MetricSerializer to ServiceCheckSerializer.
